requirements.txt
fastapi==0.115.4
httpx==0.27.2
motor==3.1.1
passlib==1.7.4
pydantic==1.10.2
pymongo==4.3.2
python-dotenv==1.0.1
python_jose==3.3.0
requests==2.32.3
uvicorn==0.32.0
email-validator==1.3.0
pytest==8.3.3   
pydantic_core
pytest-asyncio